pyproject.toml
[project]
name = "xdsec-aisec-publicshow"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=4.44.0",
    "jinja2>=3.1.4",
    "openai>=1.47.1",
    "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
    "black>=24.8.0",
    "isort>=5.13.2",
]

[tool.isort]
profile = "black"
